refactor:#114 context에 맞게 prop drilling 제거
